requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pycares==4.4.0
pycparser==2.21
pymongo==4.6.2
pyTelegramBotAPI==4.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
yarl==1.9.4
